refactor(railgun): wrap controller manager in cobra.Command
