+++ railgun/main.go
@@ -1,221 +1,7 @@
-/*
-Copyright 2021 Kong, Inc.Kong, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package main
 
-import (
-	"flag"
-	"net/http"
-	"os"
-
-	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
-	// to ensure that exec-entrypoint and run can make use of them.
-	_ "k8s.io/client-go/plugin/pkg/client/auth"
-
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/healthz"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	"github.com/kong/go-kong/kong"
-	"github.com/kong/kubernetes-ingress-controller/pkg/sendconfig"
-	konghqcomv1 "github.com/kong/kubernetes-ingress-controller/railgun/apis/configuration/v1"
-	"github.com/kong/kubernetes-ingress-controller/railgun/apis/configuration/v1alpha1"
-	configurationv1alpha1 "github.com/kong/kubernetes-ingress-controller/railgun/apis/configuration/v1alpha1"
-	configurationv1beta1 "github.com/kong/kubernetes-ingress-controller/railgun/apis/configuration/v1beta1"
-	"github.com/kong/kubernetes-ingress-controller/railgun/controllers"
-	kongctrl "github.com/kong/kubernetes-ingress-controller/railgun/controllers/configuration"
-	//+kubebuilder:scaffold:imports
-)
-
-//go:generate go run github.com/kong/kubernetes-ingress-controller/railgun/cmd/generators/controllers/networking
-
-var (
-	scheme   = runtime.NewScheme()
-	setupLog = ctrl.Log.WithName("setup")
-)
-
-func init() {
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(konghqcomv1.AddToScheme(scheme))
-	utilruntime.Must(configurationv1alpha1.AddToScheme(scheme))
-	utilruntime.Must(configurationv1beta1.AddToScheme(scheme))
-	//+kubebuilder:scaffold:scheme
-}
+import "github.com/kong/kubernetes-ingress-controller/railgun/cmd/rootcmd"
 
 func main() {
-	var metricsAddr string
-	var enableLeaderElection bool
-	var probeAddr string
-	flag.StringVar(&metricsAddr, "metrics-bind-address", ":8080", "The address the metric endpoint binds to.")
-	flag.StringVar(&probeAddr, "health-probe-bind-address", ":8081", "The address the probe endpoint binds to.")
-	flag.BoolVar(&enableLeaderElection, "leader-elect", false,
-		"Enable leader election for controller manager. "+
-			"Enabling this will ensure there is only one active controller manager.")
-
-	var kongURL string
-	var filterTag string
-	var concurrency int
-	var secretName string
-	var secretNamespace string
-	flag.StringVar(&kongURL, "kong-url", "http://localhost:8001", "TODO")
-	flag.StringVar(&filterTag, "kong-filter-tag", "managed-by-railgun", "TODO")
-	flag.IntVar(&concurrency, "kong-concurrency", 10, "TODO")
-	flag.StringVar(&secretName, "secret-name", "kong-config", "TODO")
-	flag.StringVar(&secretNamespace, "secret-namespace", controllers.DefaultNamespace, "TODO")
-
-	opts := zap.Options{
-		Development: true,
-	}
-	opts.BindFlags(flag.CommandLine)
-	flag.Parse()
-
-	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&opts)))
-
-	// TODO: we might want to change how this works in the future, rather than just assuming the default ns
-	if v := os.Getenv(controllers.CtrlNamespaceEnv); v == "" {
-		os.Setenv(controllers.CtrlNamespaceEnv, controllers.DefaultNamespace)
-	}
-
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme:                 scheme,
-		MetricsBindAddress:     metricsAddr,
-		Port:                   9443,
-		HealthProbeBindAddress: probeAddr,
-		LeaderElection:         enableLeaderElection,
-		LeaderElectionID:       "5b374a9e.konghq.com",
-	})
-	if err != nil {
-		setupLog.Error(err, "unable to start manager")
-		os.Exit(1)
-	}
-
-	/* TODO: re-enable once fixed
-	if err = (&kongctrl.KongIngressReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("KongIngress"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "KongIngress")
-		os.Exit(1)
-	}
-	if err = (&kongctrl.KongClusterPluginReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("KongClusterPlugin"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "KongClusterPlugin")
-		os.Exit(1)
-	}
-	if err = (&kongctrl.KongPluginReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("KongPlugin"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "KongPlugin")
-		os.Exit(1)
-	}
-	if err = (&kongctrl.KongConsumerReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("KongConsumer"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "KongConsumer")
-		os.Exit(1)
-	}
-	*/
-
-	kongClient, err := kong.NewClient(&kongURL, http.DefaultClient)
-	if err != nil {
-		setupLog.Error(err, "unable to create kongClient")
-		os.Exit(1)
-	}
-
-	if err = (&kongctrl.SecretReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("Secret"),
-		Scheme: mgr.GetScheme(),
-		Params: kongctrl.SecretReconcilerParams{
-			WatchName:      secretName,
-			WatchNamespace: secretNamespace,
-			KongConfig: sendconfig.Kong{
-				URL:         kongURL,
-				FilterTags:  []string{filterTag},
-				Concurrency: concurrency,
-				Client:      kongClient,
-			},
-		},
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "Secret")
-		os.Exit(1)
-	}
-
-	// TODO - we've got a couple places in here and below where we "short circuit" controllers if the relevant API isn't available.
-	// This is convenient for testing, but maintainers should reconsider this before we release KIC 2.0.
-	// SEE: https://github.com/Kong/kubernetes-ingress-controller/issues/1101
-	if err := kongctrl.SetupIngressControllers(mgr); err != nil {
-		setupLog.Error(err, "unable to create controllers", "controllers", "Ingress")
-		os.Exit(1)
-	}
-
-	// TODO - similar to above, we're short circuiting here. It's convenient, but let's discuss if this is what we want ultimately.
-	// SEE: https://github.com/Kong/kubernetes-ingress-controller/issues/1101
-	udpIngressAvailable, err := kongctrl.IsAPIAvailable(mgr, &v1alpha1.UDPIngress{})
-	if !udpIngressAvailable {
-		setupLog.Error(err, "API configuration.konghq.com/v1alpha1/UDPIngress is not available, skipping controller")
-	} else {
-		if err = (&kongctrl.KongV1Alpha1UDPIngressReconciler{
-			Client: mgr.GetClient(),
-			Log:    ctrl.Log.WithName("controllers").WithName("UDPIngress"),
-			Scheme: mgr.GetScheme(),
-		}).SetupWithManager(mgr); err != nil {
-			setupLog.Error(err, "unable to create controller", "controller", "UDPIngress")
-			os.Exit(1)
-		}
-	}
-
-	tcpIngressAvailable, err := kongctrl.IsAPIAvailable(mgr, &configurationv1beta1.TCPIngress{})
-	if !tcpIngressAvailable {
-		setupLog.Error(err, "API configuration.konghq.com/v1alpha1/TCPIngress is not available, skipping controller")
-	} else {
-		if err = (&kongctrl.KongV1Beta1TCPIngressReconciler{
-			Client: mgr.GetClient(),
-			Log:    ctrl.Log.WithName("controllers").WithName("TCPIngress"),
-			Scheme: mgr.GetScheme(),
-		}).SetupWithManager(mgr); err != nil {
-			setupLog.Error(err, "unable to create controller", "controller", "TCPIngress")
-			os.Exit(1)
-		}
-	}
-	//+kubebuilder:scaffold:builder
-
-	if err := mgr.AddHealthzCheck("health", healthz.Ping); err != nil {
-		setupLog.Error(err, "unable to set up health check")
-		os.Exit(1)
-	}
-	if err := mgr.AddReadyzCheck("check", healthz.Ping); err != nil {
-		setupLog.Error(err, "unable to set up ready check")
-		os.Exit(1)
-	}
-
-	setupLog.Info("starting manager")
-	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
-		setupLog.Error(err, "problem running manager")
-		os.Exit(1)
-	}
+	rootcmd.Execute()
 }
